requirements.txt
flask==2.0.1
rpi_ws281x==4.3.4
adafruit-circuitpython-neopixel==6.3.9
python-escpos==3.0a8
numpy==1.21.2
python-dotenv==0.19.0
gunicorn==20.1.0
textblob==0.15.3 